pyproject.toml
[project]
name = "masterblaster-classifier-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "backoff>=2.2.1",
    "black>=25.1.0",
    "codebase-to-text>=1.0.7",
    "datasets>=3.3.2",
    "kagglehub>=0.3.10",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "openai>=1.65.2",
    "openai-manager>=1.0.0",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "python-levenshtein>=0.26.1",
    "ruff>=0.9.9",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "typing-extensions>=4.12.2",
]
